Moving integration tests to their own crate
