+++ engine/Cargo.toml
@@ -51,14 +51,3 @@ strum_macros = "0.20.1"
 version = "1.0.39"
 features = [ "full" ]
 #features = [ "full", "extra-traits" ]
-
-[dev-dependencies]
-trybuild = "1.0"
-test-env-log = "0.2.2"
-env_logger = "0.9.0"
-rust_info = "0.3.1"
-cc = "1.0"
-# This is necessary for building the projects created
-# by the trybuild test system...
-autocxx = { path=".." }
-link-cplusplus = "1.0"
