+++ integration-tests/Cargo.toml
@@ -1,0 +1,28 @@
+[package]
+name = "autocxx-integration-tests"
+version = "0.1.0"
+edition = "2018"
+
+[dev-dependencies]
+proc-macro2 = "1.0.11"
+trybuild = "1.0"
+test-env-log = "0.2.2"
+env_logger = "0.9.0"
+rust_info = "0.3.1"
+cc = "1.0"
+quote = "1.0"
+once_cell = "1.7"
+autocxx-engine = { version="=0.12.0", path="../engine", features = ["build"] }
+# This is necessary for building the projects created
+# by the trybuild test system...
+autocxx = { path=".." }
+link-cplusplus = "1.0"
+tempfile = "3.1"
+indoc = "1.0"
+log = "0.4"
+cxx = "1.0.54"
+
+[dependencies.syn]
+version = "1.0.39"
+features = [ "full" ]
+#features = [ "full", "extra-traits" ]
